DjangoTuts/requirements.txt
django==4.2.8openai==1.3.7
djangorestframework==3.14.0
django-environ==0.11.2
